feat: Move ServerConfigRepositoryImpl to data module (Phase 9 start)
First step of Phase 9: move pure Kotlin repository implementations
to the shared data module.

- ServerConfigRepositoryImpl → data/src/commonMain/kotlin/
- Inject serverConfigKey via constructor (was global APP_CONFIG ref)
- Remove duplicate ServerConfigRepository interface from config/
  (domain/ already has the canonical interface)
- Update all imports (4 source files + 1 test file)
- DI wiring passes APP_CONFIG.serverConfigKey at construction time
