+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/door/DoorRepositoryImpl.kt
@@ -19,12 +19,12 @@ package com.chriscartland.garage.door
 
 import co.touchlab.kermit.Logger
 import com.chriscartland.garage.config.APP_CONFIG
-import com.chriscartland.garage.config.ServerConfigRepository
 import com.chriscartland.garage.data.LocalDoorDataSource
 import com.chriscartland.garage.data.NetworkDoorDataSource
 import com.chriscartland.garage.domain.model.DoorEvent
 import com.chriscartland.garage.domain.model.DoorPosition
 import com.chriscartland.garage.domain.repository.DoorRepository
+import com.chriscartland.garage.domain.repository.ServerConfigRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
